+++ .gitignore
@@ -1,5 +1,3 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
 # dependencies
 /node_modules
 /.pnp
@@ -17,6 +15,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.env.backend
 
 npm-debug.log*
 yarn-debug.log*
