--- craco.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    style: {
-      postOptions: {
-        plugins: [require("tailwindcss"), require("autoprefixer")],
-      },
-    },
-  };
