+++ index.html
@@ -5,9 +5,9 @@
   <meta charset="utf-8" />
   <meta name="description" content="Edgy - Best Social Media Growth Company">
   <meta name="keywords" content="Edgy, social media agency, social media growth, instagram growth, social media growth">
-  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+  <link rel="icon" href="/favicon.ico" />
+  <link href="/src/core-ui/index.css" rel="stylesheet">
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
   <link rel="preload" href="https://www.google.com/recaptcha/api.js" as="script">
   <script src="https://www.google.com/recaptcha/api.js" async defer></script>
   <title>Edgy - #1 Instagram Growth Agency</title>
@@ -15,6 +15,7 @@
 
 <body>
   <main id="root" class="m-0 p-0"></main>
+  <script type="module" src="/src/index.jsx"></script>
 </body>
 
 </html>
