+++ src/airtable/airtable.js
@@ -1,8 +1,7 @@
-//Airtable
 import Airtable from "airtable";
 
-const apiKey = process.env.REACT_APP_AIRTABLE_KEY;
-const baseId = process.env.REACT_APP_AIRTABLE_ID;
+const apiKey = import.meta.env.VITE_AIRTABLE_KEY;
+const baseId = import.meta.env.VITE_AIRTABLE_ID;
 
 const edgyBase = new Airtable({ apiKey: apiKey }).base(baseId);
