+++ src/index.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App";
-//Styles
 import "./core-ui/index.css";
-import reportWebVitals from "./reportWebVitals";
 
 const container = document.getElementById("root");
 const root = createRoot(container);
@@ -12,5 +10,3 @@ root.render(
     <App />
   </React.StrictMode>
 );
-
-reportWebVitals();
