+++ src/routes/faq/FaqList.jsx
