+++ src/routes/landing/Partners.jsx
