+++ src/routes/pricing/PricingFaq.jsx
