+++ src/routes/terms/Terms.jsx
